Falcon 9 data collected from space x website using API
The data is finally stored in a data frame that contains data specific to falcon 9 launch. A few new columns were extracted from the coded data in rockets, launchpad, cores, and payloads columns. Missing values in the payload column were resolved by replacing them using mean mass of the payloads. 

As an outcome, the data was exported to a csv file.