EDA performed on falcon 9 launch dataset
EDA was performed on the data set that was generated in in the last section where data was collected from spaceX website using API and some cleaning as well as normalization operations were performed. The data contained launched events with falcon 9 booster and related parameters. In this notebook, successful events were observed and landing success of stage 2 is classified as 0 for failure and 1 for success. The list "class" was created and added to the dataframe (containing complete flacon 9 data) as a column.